refactor theme toggler component
