+++ packages/front-end/components/Layout/TopNav.tsx
@@ -1,4 +1,4 @@
-import { FC, useCallback, useEffect, useState } from "react";
+import { FC, useState } from "react";
 import { useWatching } from "../../services/WatchProvider";
 import useGlobalMenu from "../../services/useGlobalMenu";
 import { useForm } from "react-hook-form";
@@ -11,8 +11,6 @@ import {
   FaBell,
   FaBuilding,
   FaExclamationTriangle,
-  FaLightbulb,
-  FaMoon,
 } from "react-icons/fa";
 import Link from "next/link";
 import Avatar from "../Avatar";
@@ -27,21 +25,7 @@ import Head from "next/head";
 import useStripeSubscription from "../../hooks/useStripeSubscription";
 import UpgradeModal from "../Settings/UpgradeModal";
 import Tooltip from "../Tooltip";
-
-const HalfCircleIcon: FC = () => (
-  <svg
-    width="15"
-    height="15"
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M12 22.5C14.7848 22.5 17.4555 21.3938 19.4246 19.4246C21.3938 17.4555 22.5 14.7848 22.5 12C22.5 9.21523 21.3938 6.54451 19.4246 4.57538C17.4555 2.60625 14.7848 1.5 12 1.5V22.5ZM12 24C8.8174 24 5.76516 22.7357 3.51472 20.4853C1.26428 18.2348 0 15.1826 0 12C0 8.8174 1.26428 5.76516 3.51472 3.51472C5.76516 1.26428 8.8174 0 12 0C15.1826 0 18.2348 1.26428 20.4853 3.51472C22.7357 5.76516 24 8.8174 24 12C24 15.1826 22.7357 18.2348 20.4853 20.4853C18.2348 22.7357 15.1826 24 12 24Z"
-      fill="currentColor"
-    />
-  </svg>
-);
+import { ThemeToggler } from "./ThemeToggler";
 
 const TopNav: FC<{
   toggleLeftMenu?: () => void;
@@ -49,56 +33,6 @@ const TopNav: FC<{
   showNotices?: boolean;
 }> = ({ toggleLeftMenu, pageTitle, showNotices }) => {
   const router = useRouter();
-
-  // region UI theme
-
-  const [themeDropdownOpen, setThemeDropdownOpen] = useState(false);
-  const [activeTheme, setActiveTheme] = useState<"system" | "dark" | "light">(
-    "system"
-  );
-
-  useEffect(() => {
-    try {
-      const themeFromStorage = localStorage.getItem("gb_ui_theme") as
-        | "system"
-        | "dark"
-        | "light"
-        | undefined;
-
-      if (themeFromStorage === "light" || themeFromStorage === "dark") {
-        document.documentElement.className = `theme--${themeFromStorage}`;
-        setActiveTheme(themeFromStorage);
-      }
-    } catch (e) {
-      // Nothing
-    }
-  }, []);
-
-  const handleThemeChange = useCallback(
-    (theme: "dark" | "light" | "system") => {
-      setActiveTheme(theme);
-
-      document.documentElement.className = `theme--${theme}`;
-
-      setThemeDropdownOpen(false);
-
-      switch (theme) {
-        case "dark":
-          localStorage.setItem("gb_ui_theme", "dark");
-          break;
-        case "light":
-          localStorage.setItem("gb_ui_theme", "light");
-          break;
-        case "system":
-          localStorage.removeItem("gb_ui_theme");
-          break;
-      }
-    },
-    []
-  );
-
-  // endregion UI theme
-
   const [userDropdownOpen, setUserDropdownOpen] = useState(false);
   const [orgDropdownOpen, setOrgDropdownOpen] = useState(false);
   const { watchedExperiments, watchedFeatures } = useWatching();
@@ -107,7 +41,6 @@ const TopNav: FC<{
   const [upgradeModal, setUpgradeModal] = useState(false);
   useGlobalMenu(".top-nav-user-menu", () => setUserDropdownOpen(false));
   useGlobalMenu(".top-nav-org-menu", () => setOrgDropdownOpen(false));
-  useGlobalMenu(".top-nav-theme-menu", () => setThemeDropdownOpen(false));
 
   const {
     showSeatOverageBanner,
@@ -209,55 +142,7 @@ const TopNav: FC<{
 
         <div style={{ flex: 1 }} />
 
-        <div className="dropdown top-nav-theme-menu">
-          <div
-            className={`nav-link `}
-            onClick={(e) => {
-              e.preventDefault();
-              setThemeDropdownOpen(!themeDropdownOpen);
-            }}
-            style={{ cursor: "pointer" }}
-          >
-            <button className="btn mr-1 text-secondary">
-              <FaMoon />
-            </button>
-          </div>
-
-          <div
-            className={clsx("dropdown-menu dropdown-menu-right", {
-              show: themeDropdownOpen,
-            })}
-          >
-            <button
-              onClick={() => handleThemeChange("system")}
-              className={clsx("dropdown-item", {
-                primary: activeTheme !== "light" && activeTheme !== "dark",
-              })}
-            >
-              <span className="mr-2">
-                <HalfCircleIcon />
-              </span>
-              System Default
-            </button>
-            <button
-              onClick={() => handleThemeChange("light")}
-              className={clsx("dropdown-item", {
-                primary: activeTheme === "light",
-              })}
-            >
-              <FaLightbulb className="mr-1" /> Light mode
-            </button>
-
-            <button
-              onClick={() => handleThemeChange("dark")}
-              className={clsx("dropdown-item", {
-                primary: activeTheme === "dark",
-              })}
-            >
-              <FaMoon className="mr-1" /> Dark mode
-            </button>
-          </div>
-        </div>
+        <ThemeToggler />
 
         {showNotices && (
           <>
